requirements.txt
aiofiles==24.1.0
anyio==4.9.0
asyncio==3.4.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dotenv==0.9.9
Flask==3.1.1
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.52.0
priority==2.0.0
pyee==13.0.0
python-dotenv==1.1.0
python-telegram-bot==22.1
Quart==0.20.0
quart-wtf==0.0.1
requests==2.32.3
selectolax==0.3.29
sniffio==1.3.1
telegram_text==0.2.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
